Remove print_recent_activity functions from mods
Totara has a different way of presenting recent activity.
The Moodle way has been deprecated and the existing of any
{mod}_print_recent_activity causes debug notices.
Harmless but annoying.

The Kalvidres and Kalvidassign modules both have a
_print_recent_activity function within their lib.php.
However the function simply returns false, and does nothing.

This commit removes the function so that the debugging notices
are not shown.

This only affects sites that have turned debugging on.
Not recommended for production sites.