Fix handling of preview for embedding outside of editors
The LTI launch scripts within each plugin all use a central function
when generating the launch URL.
That function expects the caller to proivide editor that is the launch
is being used for.
However due to design if the editor is not provided it guesses.
Within Totara at least this leads to the atto preview being used for all
lti launch calls, which does not work unless used within Atto as it
looks for JS API published by Atto.

To fix this I changed the operation of the central function.
The editor argument now expects to be given one of three values, each
with a meaning:
* String - the name of the editor I am being used within e.g. "atto"
* False - not used within an editor
* True - used within an editor, but we don't know which, so guess.

I haven't looked into whether the functionality works within Moodle but
the code in this area seems confused, and I would guess in Moodle it
would work for TinyMCE but not for Atto.

This is required for Totara, but should be tested for Moodle.

To reproduce:
* Create a new Kaltura Video Assignment
* Attempt to select media from Kaltura
* Confirm you can complete the selection and see the media resource
* Also affects the Kaltura Video resource and Video presentation plugins
* Repeat in all editors